requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
CacheControl==0.12.11
cachetools==5.2.0
celery==5.2.6
certifi==2022.6.15
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.13
django-autoslug==1.9.8
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-cors-headers==3.11.0
django-phonenumber-field==6.1.0
django-qr-code==3.0.0
django-rest-framework==0.1.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
fcm-django==1.0.12
firebase-admin==5.2.0
flower==1.0.0
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.20.4
humanize==4.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.4
msgpack==1.0.3
packaging==21.3
phonenumbers==8.12.44
Pillow==9.0.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
proto-plus==1.22.0
protobuf==3.20.1
psycopg==3.0.9
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
qrcode==7.3.1
redis==4.2.2
requests==2.28.1
rsa==4.9
segno==1.4.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
Twisted==22.2.0
txaio==22.2.1
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
vonage==3.0.0
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.14.1
zope.interface==5.4.0
